# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Muttley API
  description: >-
    API REST do Sistema Muttley para Gestão de Eventos Acadêmicos, Inscrições,
    Emissão e Validação de Certificados com Assinatura Visual e Gamificação por Medalhas.
    Desenvolvido para a FATEC Zona Leste.
  version: 0.1.0-alpha
  contact:
    name: Equipe Muttley - FATEC Zona Leste
    url: https://github.com/andrelamego/Backend-Muttley
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8083
    description: Servidor Local de Desenvolvimento
...
paths:
...
  /api/pessoa/dados-cadastro/{token}:
    get:
      tags:
        - Pessoas e Perfis
      summary: Consultar dados para conclusão de cadastro com convite
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dados retornados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterInfo'
        '404':
          description: Não encontrada
...
  /api/admin/pessoas:
    get:
      tags:
        - Pessoas e Perfis
      summary: Listar todas as pessoas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de pessoas
    post:
      tags:
        - Pessoas e Perfis
      summary: Cadastrar pessoa
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizacaoPessoa'
      responses:
        '201':
          description: Criada com sucesso
  /api/admin/pessoas/{id}:
    get:
      tags:
        - Pessoas e Perfis
      summary: Buscar pessoa por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Pessoa encontrada
    put:
      tags:
        - Pessoas e Perfis
      summary: Atualizar pessoa
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizacaoPessoa'
      responses:
        '200':
          description: Atualizada com sucesso
    delete:
      tags:
        - Pessoas e Perfis
      summary: Excluir pessoa
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Excluída com sucesso
  /api/admin/alunos:
    get:
      tags:
        - Pessoas e Perfis
      summary: Listar alunos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de alunos
  /api/admin/professores:
    get:
      tags:
        - Pessoas e Perfis
      summary: Listar professores
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de professores
  /api/admin/palestrantes:
    get:
      tags:
        - Pessoas e Perfis
      summary: Listar palestrantes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de palestrantes
  /api/admin/organizadores:
    get:
      tags:
        - Pessoas e Perfis
      summary: Listar organizadores
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de organizadores
  /api/admin/colaboradores:
    get:
      tags:
        - Pessoas e Perfis
      summary: Listar colaboradores
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de colaboradores
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Insira o token JWT retornado no endpoint /api/auth/login.
  schemas:
...
    AtualizacaoPessoa:
      type: object
      required: [nome, email, telefone, cpf, senha]
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
          example: Maria Souza
        email:
          type: string
          format: email
          example: maria@fatec.sp.gov.br
        cpf:
          type: string
          example: '529.982.247-25'
        telefone:
          type: string
          example: '11999999999'
        senha:
          type: string
          example: SenhaForte@123
...
    RegisterInfo:
      type: object
      properties:
        nome:
          type: string
        email:
          type: string
        cpf:
          type: string